requirements.txt
streamlit==0.89.0
Pillow==8.2.0